Issue #9 Typesafe path parameter substitution
Generate an enum per resource with available path parameters as values.

TemplatedPath is now generic and type-safe: e.g., you cannot
substitute ProductResourcePathParameters instances when
BrandResourcePathParameters instances are expected without
angering the compiler :)
Resources that do not define path parameters simply expose
NoPathParameters, which triggers an exception when called.

Use a LinkedHashMultimap to store parsed sources to have
iterators that follow the ordering in which the data was added to the multimap.
It makes the build safer using different JDK (v8 and v7 was not returning the same
ordering).