Use the QueryParam annotation value as name
We used to use the parameter name instead of the annotated value
as name when generating our enum for Query parameters.

This is problematic when we are going to handle bean params as
annotated arguments, given they may be on constructors (therefore
will have a very bad name : arg0, etc.).

Instead of using the argument name, let us use the jersey annotation
value as name.