Use the QueryParam annotation value as name
We used to use the parameter name instead of the annotated value
as name when generating our enum for Query parameters.
This is problematic when we're going to handle bean params as annotated arguments
may be on constructors (therefor will have a very bad name : arg0, etc.).
So instead of using the argument name, let's use the jersey annotation value as name.