Add a module to report IT and UT code coverage
https://github.com/jacoco/jacoco/wiki/MavenMultiModule

Choosing the module with dependencies strategy (https://github.com/jacoco/jacoco/wiki/MavenMultiModule#strategy-module-with-dependencies)
as it feels the less awkward one 😇

Replaces the run-forked goal by start for jetty because it causes re-compilation of the sources and out of sync coverage report.

The coveralls plugins has also been tweaked because errors like:

    I/O operation failed: No source found for fr/vidal/oss/jax_rs_linker/model/PathParameter.java

When used on a module with generated sources the plugin can't handle these sources.
In a traditional build we should not want to measure such code but as we are testing that our annotation processor
is working as expected we need these reports 😎